requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/test.in
#
attrs==22.1.0
    # via pytest
exceptiongroup==1.0.4
    # via pytest
factory-boy==3.2.1
    # via pytest-factoryboy
faker==15.3.4
    # via factory-boy
flake8==6.0.0
    # via -r requirements/test.in
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
mccabe==0.7.0
    # via flake8
packaging==22.0
    # via pytest
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytest==7.2.0
    # via
    #   -r requirements/test.in
    #   pytest-django
    #   pytest-factoryboy
pytest-django==4.5.2
    # via -r requirements/test.in
pytest-factoryboy==2.5.1
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via pytest
typing-extensions==4.4.0
    # via pytest-factoryboy
